[Fix] 지도 스타일 수정
