+++ front/src/pages/MapPage/style.css
@@ -1,0 +1,27 @@
+.map-area {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+}
+
+/* 지도 */
+.map-area > div:first-child {
+  flex: 2;
+  width: 100%;
+  height: 100%;
+}
+
+/* 리스트 */
+.map-area > div:last-child {
+  flex: 1;
+  width: 100%;
+  height: 100%;
+}
+
+/* PC */
+@media (min-width: 768px) {
+  .map-area {
+    flex-direction: row;
+  }
+}
